Export DX & DXDefault separately
Still styleguidist complains...
