+++ styleguide.config.js
@@ -1,6 +1,114 @@
+// @format
+
+const path = require("path");
+
+const reactDocgenTypescript = require("react-docgen-typescript").withCustomConfig(
+    "./tsconfig.json"
+);
+
+const typescriptPropsParser = reactDocgenTypescript.parse;
+
 module.exports = {
-    components: 'src/**/[A-Z]*.{ts,tsx}',
-    propsParser: require('react-docgen-typescript').withCustomConfig(
-        './tsconfig.json'
-    ).parse
-}
+    title: "Data Explorer",
+    defaultExample: false,
+    propsParser: typescriptPropsParser,
+    resolver: require("react-docgen").resolver.findAllComponentDefinitions,
+    getComponentPathLine: componentPath => {
+        const toPascalCase = string =>
+            string
+                .match(/[A-Z]+/gi)
+                .map(word => word.charAt(0).toUpperCase() + word.substr(1).toLowerCase())
+                .join("");
+        const name = path.basename(componentPath, ".tsx");
+        const from = componentPath
+            //.replace(/^packages\//, "@nteract/")
+            .replace(/\.tsx?$/, "");
+        return `import ${toPascalCase(name)} from '${from}';`;
+    },
+    /*  sections: [
+        {
+          name: "Introduction",
+          content: "styleguide-components/intro.md"
+        },
+        {
+          name: "@nteract/presentational-components",
+          components: "packages/presentational-components/src/components/*.tsx"
+        },
+        {
+          name: "@nteract/outputs",
+          components: "packages/outputs/src/components/*.tsx"
+        },
+        {
+          name: "@nteract/outputs/media",
+          components: "packages/outputs/src/components/media/*.tsx",
+          content: "packages/outputs/src/components/media/index.md",
+          ignore: "packages/outputs/src/components/media/index.tsx"
+        },
+        // {
+        //   name: "@mybinder/host-cache",
+        //   components: "packages/host-cache/src/components/*.tsx"
+        // },
+        {
+          name: "@nteract/directory-listing",
+          components: "packages/directory-listing/src/components/*.tsx"
+        },
+        {
+          name: "@nteract/markdown",
+          content: "packages/markdown/examples.md"
+        },
+        {
+          name: "@nteract/mathjax",
+          content: "packages/mathjax/examples.md"
+        }
+      ], */
+    // For overriding the components styleguidist uses
+    /*  styleguideComponents: {
+        LogoRenderer: path.join(__dirname, "styleguide-components", "logo.tsx")
+      }, */
+    compilerConfig: {
+        // Allow us to use {...props}
+        objectAssign: "Object.assign",
+        transforms: {
+            // whether template strings get transpiled (we don't want it to, so that we can use the native functionality)
+            templateString: false
+        }
+    },
+    template: {
+        body: {
+            raw: ``
+        }
+    },
+    webpackConfig: {
+        node: {
+            fs: "empty",
+            child_process: "empty",
+            net: "empty",
+            canvas: "empty"
+        },
+        resolve: {
+            extensions: [".ts", ".tsx", ".js", ".jsx", ".json"]
+        },
+        externals: ["canvas"],
+        module: {
+            rules: [
+                {
+                    test: /\.tsx?$/,
+                    loader: "ts-loader",
+                    options: {
+                        compilerOptions: {
+                            strict: true,
+                            jsx: "react",
+                            composite: true
+                        },
+                        projectReferences: true,
+                        transpileOnly: true
+                    }
+                },
+                {
+                    test: /\.css$/,
+                    use: ['style-loader', 'css-loader']
+                },
+            ]
+        }
+    }
+};
