--- webpack.config.js
@@ -1,28 +1,0 @@
-const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
-
-module.exports = {
-    entry: "./src/index.tsx",
-    output: {
-        filename: "bundle.js",
-        path: __dirname + "/dist"
-    },
-
-    // Enable sourcemaps for debugging webpack"s output.
-    devtool: "source-map",
-
-    resolve: {
-        // Add ".ts" and ".tsx" as resolvable extensions.
-        extensions: [".ts", ".tsx", ".js", ".json"],
-            plugins: [new TsconfigPathsPlugin({ configFile: "./tsconfig.json" })]
-        
-    },
-
-    module: {
-        rules: [
-            { test: /\.tsx?$/, loader: "ts-loader" },
-            { enforce: "pre", test: /\.js$/, loader: "source-map-loader" },
-            { test: /\.css$/, use: ["style-loader", "css-loader"] },
-        ]
-    },
-
-};
